Refactor API requests and responses
