--- content/Bat/Bat.Api/Controllers/ApiReqModels.cs
@@ -1,2 +1,0 @@
-﻿namespace Bat.Api.Controllers;
-
