+++ content/Bat/Bat.Blazor/Bat.Blazor.App/Pages/RolesAdd.razor
@@ -42,13 +42,17 @@
 							@foreach (var claim in ClaimList)
 							{
 								<div class="form-check">
-									<input class="form-check-input" type="checkbox" id="claim-@claim.ClaimType-@claim.ClaimValue" value="@claim.ClaimType:@claim.ClaimValue">
+									<input class="form-check-input" type="checkbox" id="claim-@claim.ClaimType-@claim.ClaimValue" value="@claim.ClaimType:@claim.ClaimValue" @onchange="(()=>OnClaimChanged(claim.ClaimType, claim.ClaimValue))">
 									<label class="form-check-label" for="claim-@claim.ClaimType-@claim.ClaimValue">@claim.ClaimType:@claim.ClaimValue</label>
 								</div>
 							}
 						}
 					</form>
                 </div>
+				<div class="card-footer">
+					<button class="btn btn-primary" @onclick="(()=>BtnClickCreate())"><CIcon icon="cil-save" /> Create Role</button>
+					<button class="btn btn-secondary" @onclick="(()=>BtnClickCancel())"><CIcon icon="cil-x-circle" /> Cancel</button>
+				</div>
             </div>
         </div>
     </Authorized>
